Added all the necessary documentation to what I believe to be a sufficient level

+++ Project0/src/main/java/com/game/service/AccountService.java
@@ -7,6 +7,10 @@ import com.game.models.Message;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Services both the account repo and the message repo
+ * Possesses an account/message list for quicker access
+ */
 public class AccountService {
     List<Account> accountList;
     List<Message> messageList;
@@ -14,16 +18,29 @@ public class AccountService {
     Repository<Account, String> repo;
     MessageSQLRepo mrepo;
 
+    /**
+     * Allows service to utilize the repos
+     * @param repo account repo
+     * @param mrepo message repo
+     */
     public AccountService(Repository<Account, String> repo,Repository<Message, Integer> mrepo) {
         this.repo = repo;
         this.mrepo = (MessageSQLRepo) mrepo;
     }
 
+    /**
+     * Generates the account list
+     */
     public void boot() {
         accountList = repo.findAll();
     }
 
-    //returns true if there is an existing account with that name
+    /**
+     * Checks if there is another account with the given username
+     * Helps prevent duplicate as the username field is unique
+     * @param username username given
+     * @return false if there is an existing account with that name
+     */
     public boolean checkDuplicates(String username) {
         //checks if the name is already in list
         //returns true if it passes the check
@@ -36,7 +53,13 @@ public class AccountService {
         return true;
     }
 
-    //returns true if account is found and sets it equal to curr
+    /**
+     * Checks whether credentials match an account and sets a reference to it
+     * Generates a message list from the account information
+     * @param username given username
+     * @param password given password
+     * @return true if account is found and sets it equal to curr
+     */
     public boolean checkCredentials(String username, String password) {
         System.out.println(username+"\t"+password);
         for (Account account : accountList) {
@@ -57,7 +80,14 @@ public class AccountService {
         return false;
     }
 
-    //created for the sign up process while setting up the curr account reference
+    /**
+     * Creates account object with standard user level and adds it to the account
+     * list and repo. Also sets up a new empty message list
+     * created for the sign up process while setting up the curr account reference
+     * @param username username of the new account
+     * @param password password of the new account
+     * @return true if the account has been created
+     */
     public boolean signUp(String username, String password){
         if (checkDuplicates(username)) {
             curr = new Account(username, password);
@@ -72,6 +102,13 @@ public class AccountService {
         return false;
     }
 
+    /**
+     * Creates a new account object and adds it to the list and repo
+     * @param username username of new account
+     * @param password password of newly created account
+     * @param isadmin whether the account is admin level
+     * @return true if account has been created successfully
+     */
     public boolean createAccount(String username, String password, boolean isadmin) {
         if(checkDuplicates(username)) {
             Account temp = new Account(username, password, isadmin);
@@ -82,6 +119,10 @@ public class AccountService {
         return false;
     }
 
+    /**
+     * checks administer status of current account
+     * @return true if current user is admin level
+     */
     public boolean getIsAdminStatus(){
         return curr.isAdmin();
     }
@@ -122,6 +163,12 @@ public class AccountService {
         System.out.println("Your account has been remove");
     }
 
+    /**
+     * Finds reference to account by traversing the account list
+     * and matching the username
+     * @param username requested account's username
+     * @return account object requested
+     */
     private Account findAccount(String username){
         for (Account temp:accountList) {
             if (temp.getName().equals(username)){
@@ -139,12 +186,19 @@ public class AccountService {
         return curr;
     }
 
-    //method to update account changes in the repo
+    /**
+     * method to update account changes in the repo
+     * Called when account state changes
+     * @param account changed account
+     */
     public void save(Account account) {
         repo.save(account);
     }
 
-    //get current account information
+    /**
+     * get account information of requested user
+     * @param username username of account you wish to look up
+     */
     public void getAccountInfo(String username) {
         Account temp = findAccount(username);
         if (temp!=null) {
@@ -153,7 +207,9 @@ public class AccountService {
         }
     }
 
-    //get all account information (admin-accessible only)
+    /**
+     * get all account information (admin-accessible only)
+     */
     public void list() {
         for (Account temp:accountList) {
             System.out.println(temp.getName()+"\t"+temp.getPassword()+"\t"+temp.getBalance()+
@@ -161,33 +217,54 @@ public class AccountService {
         }
     }
 
-    //deposit credits into current account
+    /**
+     * deposit credits into current account
+     * @param deposit credits you want to deposit
+     */
     public void depositM(int deposit) {
         curr.addCredits(deposit);
         System.out.println("Your balance is now " + curr.getBalance());
         repo.update(curr);
     }
 
-    //request credits from current account
+    /**
+     * request credits from current account
+     * and prints remaining balance
+     * @param request amount you want to withdraw
+     * @return true if the transaction is successful
+     */
     public boolean spendC(int request) {
         if(curr.spendCredits(request)) {
             System.out.println("Your balance is now " + curr.getBalance());
+            repo.update(curr);
             return true;
         }
         return false;
     }
 
-    public void gift(String username, int choice) {
-        if(curr.spendCredits(choice)){
+    /**
+     * Takes credit from current account and sends it to another
+     * @param username username of desired recipient
+     * @param request amount you wish to send
+     */
+    public void gift(String username, int request) {
+        if(curr.spendCredits(request)){
             Account temp = findAccount(username);
             if (temp!=null) {
-                temp.addCredits(choice);
+                temp.addCredits(request);
+                repo.update(temp);
+                repo.update(curr);
+            }
+            else {
+                curr.addCredits(request);
             }
         }
     }
 
-    //this section access the message service
-    // prints out content of message to the user in a numbered format
+    /**
+     * prints out content of message to the user in a numbered format
+     * Shows the username of who sent the message: message
+     */
     public void readMessages() {
         int i=1;
         for (Message m:messageList) {
@@ -218,7 +295,7 @@ public class AccountService {
      */
     public void delete(int index) {
         --index;
-        if (index < messageList.size()&index>=0) {
+        if (index < messageList.size()&&index>=0) {
             Message temp = messageList.remove(index);
             mrepo.delete(temp.getId());
         }
@@ -228,23 +305,29 @@ public class AccountService {
         System.out.println("You have "+messageList.size()+" messages");
     }
 
-    //deletes by traversing through the user's messageList and getting their unique ids from the object
+    /**
+     * deletes by traversing through the user's messageList and getting their unique ids from the object
+     */
     public void deleteAll() {
         for (int i = 0; i<messageList.size(); i++) {
             delete(i);
         }
     }
 
+    /**
+     * prints out the current user's username, password, balance, and account type
+     */
     public void getAccountInfo() {
         System.out.println(curr.getName()+"\t"+curr.getPassword()+"\t"+curr.getBalance()+
                 "\t"+(curr.isAdmin()?"Admin":"Player"));
     }
 
     /**
-     * Self explanatory
+     * Changes current user's password
      * @param newPassword desired password
      */
     public void changePassword(String newPassword){
         curr.setPassword(newPassword);
+        repo.update(curr);
     }
 }
